requirements.txt
tensorflow<2.11
opencv-python==4.8.0.76
customtkinter
pillow
numpy
pandas
